requirements.txt
# Core
numpy<2.0.0
pandas
scikit-learn==1.2.0
torch==1.13.1
tensorly==0.8.1
tqdm>4.29.0
torchsummary
sympy
matplotlib
seaborn

# For Professionalism & Complexity
wandb
cupy-cuda11x
opt_einsum
hydra-core
pyyaml